Handle Schwab price splits
Schwab splits prices when no stocks from a reward are sold. Otherwise, it doesn't.
Using this as the shelve of my solution to the problem while Max works on a historical-price-based solution.

This solution enforces the user to provide full history so that the software can correctly deduce which rewards were sold (partially or fully).